commit not full

+++ src/main/java/com/demo/event/model/dto/response/GroupSummaryResponse.java
@@ -6,8 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class GroupSummaryResponse {
-    private String groupType;    // GIA_DINH, VO_CHONG...
-    private String displayName;  // "Gia đình", "Vợ/Chồng"...
-    private Integer count;       // số người trong nhóm
+    private String groupType;
+    private String displayName;
+    private Long count;
 }
-
